+++ ProCinema/ManagerVenta.h
@@ -7,7 +7,7 @@ public:
     void listarVentas();
     void modificarVenta();
     void cambiarEstadoVenta();
-    int calcularImporteTotal(int cantidadEntradas, int idFuncion, int idMembresia);
+    float calcularImporteTotal(int cantidadEntradas, int idFuncion, int idMembresia);
 
     void buscarPorId();
     void buscarPorFuncion();
